requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
streamlit
wget
imutils
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.8.2%2Bcpu-cp36-cp36m-linux_x86_64.whl
tqdm>=4.41.0
pytesseract==0.3.7

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools==4.0
# onnx>=1.8.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
# thop  # FLOPS computation
# pycocotools>=2.0  # COCO mAP
